Insumos correcciones

+++ tpfinal-backend/app/Http/Controllers/CategoriaInsumosController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\categoriaInsumo;
+use App\Models\insumo;
 use Illuminate\Http\Request;
 
 class CategoriaInsumosController extends Controller
@@ -36,9 +37,17 @@ class CategoriaInsumosController extends Controller
         return response()->json(['exito' => true, 'message' => 'Categoria actualizada exitosamente'], 200);
     }
 
-    public function delete($id){
+    public function delete($id)
+    {
         $categoria = categoriaInsumo::find($id);
-        $categoria->delete();
-        return response()->json(['exito' => true, 'message' => 'Categoria eliminada exitosamente'], 200);
+
+        //reviso si se usa
+        $enUso = insumo::where('id_categoria', $id)->exists();
+        if ($enUso) {
+            return response()->json(['message' => 'No se puede eliminar, existe un insumo que utiliza esta categoria.', 'exito' => false], 200);
+        } else {
+            $categoria->delete();
+            return response()->json(['exito' => true, 'message' => 'Categoria eliminada exitosamente'], 200);
+        }
     }
 }
